The html for the UI to add and delete new rows

Responsive to given permissions

Usage

cdr_row_editor_html(
  notes_txt = "",
  id,
  add_row_permission = F,
  del_row_permission = F
)

Arguments

notes_txt

message to send to the UI edit page

id

namespace id - corresponds to table name

add_row_permission

T or F: whether user has permission to add a new observation, i.e. the Unique ID for a row

del_row_permission

T or F: whether user has permission to delete a new observation, i.e. the Unique ID for a row

Value

ui output component

Examples

crudr:::cdr_row_editor_html('hello', 'iris', TRUE, TRUE)
#> 
#> --Running: cdr_row_editor_html()
#> 
#> --Granting permissions: add and delete rows 
#> <span style="display: inline-flex; align-items: center; font-size: 10px;">
#>   <div class="form-group shiny-input-container" style="width:180px;">
#>     <label class="control-label" id="iris-uid-label" for="iris-uid"></label>
#>     <input id="iris-uid" type="text" class="shiny-input-text form-control" value="" placeholder="Enter Unique ID"/>
#>   </div>
#>   <button id="iris-create_row_btn" type="button" class="btn btn-default action-button" text="Create Row" style="margin-left: 15px;">Create Row</button>
#>   <button id="iris-delete_row_btn" type="button" class="btn btn-default action-button" text="Delete Row" style="margin-left: 15px;">Delete Row</button>
#>   <span style="color:red; font-size: 130%; margin-left: 15px;">hello</span>
#> </span>